+++ R/geo.R
@@ -66,6 +66,19 @@ e_geo_3d_ <- function(e, serie = NULL, color = NULL, type = "world", rm_x = TRUE
     e$x$opts$geo3D <- series
   else
     e$x$opts$baseOption$geo3D <- series
+
+  if(type == "world"){
+    # add dependency
+    path <- system.file("htmlwidgets/lib/echarts-4.8.0", package = "echarts4r")
+    dep <- htmltools::htmlDependency(
+      name = "echarts-world",
+      version = "1.0.0",
+      src = c(file = path),
+      script = "world.js"
+    )
+
+    e$dependencies <- append(e$dependencies, list(dep))
+  }
   
   e
 }
@@ -109,5 +122,19 @@ e_geo <- function(e, map = "world", ...){
     e$x$opts$geo <- opts
   else
     e$x$opts$baseOption$geo <- opts
+
+  if(map == "world"){
+    # add dependency
+    path <- system.file("htmlwidgets/lib/echarts-4.8.0", package = "echarts4r")
+    dep <- htmltools::htmlDependency(
+      name = "echarts-world",
+      version = "1.0.0",
+      src = c(file = path),
+      script = "world.js"
+    )
+
+    e$dependencies <- append(e$dependencies, list(dep))
+  }
+
   e
 }