+++ NEWS.md
@@ -10,6 +10,7 @@ Dynamic dependencies:
 - mapbox
 - wordcloud
 - liquidfill
+- world.js (underlying world map for e_geo and e_map)
 
 This greatly improves the performances of echarts4r which now loads much faster.